+++ poseidon-merkle/benches/zk.rs
@@ -8,6 +8,7 @@
 
 use criterion::{Criterion, criterion_group, criterion_main};
 use dusk_plonk::prelude::*;
+use dusk_poseidon::{Domain, Hash};
 use poseidon_merkle::zk::opening_gadget;
 use poseidon_merkle::{Item, Opening, Tree};
 use rand::rngs::StdRng;
@@ -16,17 +17,16 @@ use rand::{RngCore, SeedableRng};
 // set max circuit size to 2^13 gates
 const CAPACITY: usize = 16;
 
-// set height and arity of the poseidon merkle tree
+// set height of the poseidon merkle tree
 const HEIGHT: usize = 17;
-const ARITY: usize = 4;
 
-type PoseidonTree = Tree<(), HEIGHT, ARITY>;
+type PoseidonTree = Tree<(), HEIGHT>;
 type PoseidonItem = Item<()>;
 
 // Create a circuit for the opening
 #[derive(Debug, Clone, Copy, Hash, Eq, PartialEq)]
 struct OpeningCircuit {
-    opening: Opening<(), HEIGHT, ARITY>,
+    opening: Opening<(), HEIGHT>,
     leaf: PoseidonItem,
 }
 
@@ -48,10 +48,7 @@ impl Default for OpeningCircuit {
 
 impl OpeningCircuit {
     /// Create a new OpeningCircuit
-    pub fn new(
-        opening: Opening<(), HEIGHT, ARITY>,
-        leaf: PoseidonItem,
-    ) -> Self {
+    pub fn new(opening: Opening<(), HEIGHT>, leaf: PoseidonItem) -> Self {
         Self { opening, leaf }
     }
 }
@@ -88,7 +85,7 @@ fn bench_zk(c: &mut Criterion) {
     for _ in 0..100 {
         let pos = rng.next_u64() % u32::MAX as u64;
         let leaf = PoseidonItem {
-            hash: dusk_poseidon::sponge::hash(&[pos.into()]),
+            hash: Hash::digest(Domain::Other, &[pos.into()])[0],
             data: (),
         };
         tree.insert(pos, leaf);
@@ -97,7 +94,7 @@ fn bench_zk(c: &mut Criterion) {
     // insert new leaf in the tree at random position to create opening
     let pos = rng.next_u64() % u32::MAX as u64;
     let leaf = PoseidonItem {
-        hash: dusk_poseidon::sponge::hash(&[pos.into()]),
+        hash: Hash::digest(Domain::Other, &[pos.into()])[0],
         data: (),
     };
     tree.insert(pos, leaf);